Housekeeping - Rename purposes fixture builder
We went to use this expecting it to contain just purposes, but found it was actually returning licence version purposes.

So, rename it to make this clearer for future use.
