+++ cypress/support/fixture-builder/licence-version-purposes.js
