+++ cypress/support/scenarios/one-return-requirement-two-points.js
@@ -1,6 +1,6 @@
 import licence from '../fixture-builder/licence.js'
+import licenceVersionPurposes from '../fixture-builder/licence-version-purposes.js'
 import points from '../fixture-builder/points.js'
-import purposes from '../fixture-builder/purposes.js'
 import returnRequirements from '../fixture-builder/return-requirements.js'
 import returnRequirementPoints from '../fixture-builder/return-requirement-points.js'
 import returnVersion from '../fixture-builder/return-version.js'
@@ -9,7 +9,7 @@ export default function basicLicenceOneReturnRequirementsWithTwoPoints () {
   const dataModel = {
     ...licence(),
     ...points(2),
-    ...purposes(2),
+    ...licenceVersionPurposes(2),
     ...returnVersion(),
     ...returnRequirements(),
     ...returnRequirementPoints(2)
